pyproject.toml
[project]
name = "indaba"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.12.0",
    "groq>=0.31.1",
    "pypdf2>=3.0.1",
    "sentence-transformers>=5.1.0",
    "streamlit>=1.49.1",
]
